refactor(html-theme): Extract smaller, easier to read methods from `HTMLTheme.onBuild`
