+++ src/themes/html/searchIndex.ts
